Add versioning.
Version consists of [major].[minor].[fix] numbers.  There is also a separate
ever-increasing 'so version'.  If a release does not: alter behaviour, adds
a feature or breaks API/ABI, only [fix] should be increased.  If the release
changes the behaviour, adds a feature or breaks API/ABI [minor] should be
increased.  It the development completes for a certain version then [major]
is increased.  If API/ABI breaks (regardless of other changes), 'so version'
is increased.

The file name is assembled as:
libampache_browser_[major].so.[so version].[minor].[fix].