Do not show the main window, leave the responsibility to the client.
The window was displayed as standalone/top-level (since it has no parent)
and then re-drawn as child window when client set the parent.  It is thus
better not to call show() on the main window here in the library and leave
that to the client.  Thanks to John Lindgren for suggestion.  Fixes #5.