Windows: Fix caching for non-ASCII profile paths; fix build failure with Qt 6.
If path for cache directory contained non-ASCII characters then the
cache was not being created.  This could happen in case user name
contained such characters or a windows user profile related environment
variable pointed to such path.

Build on Windows with Qt 6 was failing because UNICODE preprocessor
constant is set in such scenario and it caused that Unicode versions
of some windows functions were build instead of ANSI.  The interfacing
code was not compatible with Unicode versions of windows functions.
This was fixed by using Unicode version only and adapting the interfacing
code.