Add missing and remove redundant includes.
On MacOS 10.10 the build fails due to a missing header stdlib in
os_paths.cc and header errno in filesystem.cc, although it compiles
successfully on Linux and Windows.

For each file, headers for symbols which are directly used in the file
are now included.

Some redundant includes were identified, those are now removed.

Fixes #8.